AC-5449 Density test: inreased multipliers
New pipelines: density and density_aws
New test: test_density_with_increased_multipliers

+++ tests_integration/lib/infra_providers.py
@@ -622,7 +622,8 @@ class AwsProvider(InfraProvider):
         # FIXME local_exec_live does not work
         utils.local_exec(['bash', 'aws-kd-deploy/cluster/aws-kd-down.sh'],
                          env=self.env)
-        os.remove(self._inventory)
+        if self._inventory:
+            os.remove(self._inventory)
 
     def power_off(self, host):
         LOG.debug("VM Power Off: '{}'".format(host))
